Move examples directory
Move examples directory from `log/examples` to `log/_examples`.
Directories and files starting with an underscore are ignored by the go
toolchain unless explicitly listed, in particular no empty godocs are
generated for the _examples directory.

However, we still want to ensure on CI that _examples build so, add
special make target `build-examples` and have `all` depend on it.