WIP: rework julia's health split a little.
experimental to see what it looks like.

Add an interface for HTTPServer and GRPCServer to get the health data
instead of each storing it in duplicate.

Remove ServerHTTP and RegisterWith from Server. The client can just call
s.HTTP.ServeHTTP or s.GRPC.RegisterWith, which is perhaps a little more
obvious what is happening too.

Put the health data in a separate private struct and embed that in
Server to expose the SetReady method.

This split allows a user of the package to supply their own
implementation of the Service interface and construct the HTTPServer
and/or GRPCServer with that, allowing a more complicated arrangement to
be plugged in, rather than just a bool (i.e. the custom version could
perform more extensive Ready checks dynamically.

Version continues to be exported so a custom implementation of Ready can
still use Version as that is not likely to need to change.