health: Add HTTPServer.RegisterWith method
Add the `RegisterWith` method to `HTTPServer` to register all three
health paths with a mux. The mux needs to implement the `health.Router`
interface which `http.ServeMux` does. `chi` also does, for example.
Unfortunately `gorilla/mux` does not (it has a return value).

Reimplement `ServeHTTP` in with `RegisterWith` on a `http.ServeMux` to
remove the code duplication that would have otherwise existed.

Adjust the `TestHTTPMethodNotAllowed` test as the order of checking
paths and methods have changed.

Also rename `TestRegisterWith` to `TestGRPCRegisterWith` since there is
now another `RegisterWith`.