health: Add IsHealthEndpoint function for OpenCensus
Add an `IsHealthEndpoint()` function that can be used with the
OpenCensus `ochttp` plugin for collecting http server stats and
injecting trace headers.

There is little point doing tracing on the health check endpoints, and
it can add a lot of noise to logs if you do as they are triggered
frequently.

The function can be used in the creation of `ochttp.Handler` as
described in the godoc comment.