health: Add OpenCensus provider for health state
Add an `ochealth` package that adds a producer to the OpenCensus
producer global manager so that a health server's state can be exported
as metrics.

This is implemented as a separate package so that users of the health
package do not have a dependency on OpenCensus unless they explicitly
use this package.